+++ hata/discord/webhook/utils.py
@@ -3,9 +3,6 @@ __all__ = ('create_partial_webhook_from_id', )
 from ...backend.export import export
 
 from ..core import USERS
-from ..user import ZEROUSER
-
-from ..bases import ICON_TYPE_NONE
 
 from .preinstanced import WebhookType
 from .webhook import Webhook
@@ -34,19 +31,8 @@ def create_partial_webhook_from_id(webhook_id, token, type_=WebhookType.bot, cha
     try:
         webhook = USERS[webhook_id]
     except KeyError:
-        webhook = object.__new__(Webhook)
-        
-        webhook.id = webhook_id
-        
-        webhook.name = ''
-        webhook.discriminator = 0
-        webhook.application_id = 0
-        webhook.avatar_hash = 0
-        webhook.avatar_type = ICON_TYPE_NONE
-        
-        webhook.user = ZEROUSER
+        webhook = Webhook._create_empty(webhook_id)
         webhook.channel = channel
-        
         webhook.type = type_
         
         USERS[webhook_id] = webhook